switch to jinja2 templates for dinamic rendering

+++ models/RiegoModel.py
@@ -5,7 +5,7 @@ from datetime import datetime
 class RiegoConfig(BaseModel):
     time: str  # HH:MM
     duration: int  # Time on
-    min_temp: int
-    max_temp: int
+    min_temp: float
+    max_temp: float
 
     timestamp: datetime
